+++ Changelog.md
@@ -1,3 +1,37 @@
+# 0.3.0
+
+*New Features*
+- Use "move" wall restriction for cover, which more closely corresponds to typical physics for cover. Note that limited move walls (terrain walls) are possible, which would provide cover beyond the first wall.
+- Consider tiles for visibility and cover if Levels module is active.
+  - For Points or Area2d algorithms, transparent tile pixels are ignored.
+  - For Area3d algorithm, drawings can be used to create holes in tiles. See Drawing config to set an ellipse, polygon, or rectangle as a hole.
+- Add ability for tokens to ignore cover. (Closes issue #3.)
+  - DND5e: Add an actor special feat menu to ignore certain levels of cover. This is the same as, and compatible with, Simbul's Cover Calculator. The flag `dnd5e.helpersIgnoreCover` stores this value.
+  - Non-DND5e: Adds a flag, `tokenvisibility.ignoreCover`
+  - Can set whether a token ignores cover using the `token.ignoresCover` property.
+- Add "Cover Debug Tester" macro that turns on visualization of the cover algorithm.
+- Add setting to count live tokens as cover
+- Add setting to count dead tokens as cover or count half the height of dead tokens as cover. (Closes issue #2.)
+
+*Bug Fixes*
+- Fix for area algorithms where viewer and target tokens overlap.
+- Improved handling of visibility and cover when Levels is active.
+- Address issue #6 (warning message re los mapping)
+- Possible fix for issue #5 (wall is over the tokens)
+- Possible fix for issue #4 (tokens visible beyond LOS range)
+- Improved handling of terrain walls for Area3d algorithm.
+
+*Code improvements*
+- Refactor the 3d points representation for Foundry tokens, tiles, drawings, walls.
+- Add handling for terrain walls for Area3d algorithm.
+- Add handling for tiles for Area3d algorithm.
+- Add handling for tokens for Area3d algorithm.
+- Add settings for using tokens (live or dead) for cover (all algorithms).
+- Add tile handling for cover points algorithms. Transparent tile pixels do not provide cover, as expected for Levels holes.
+- Improved handling of perspective transform for different Foundry objects.
+- Cache the LOS value for a given point as is done in base Foundry.
+- Cache settings
+
 # 0.2.1
 Fix for `_hasLOS` not initialized error, which occurs sometimes with Levels installed.